+++ stat_bsd.go
@@ -11,8 +11,12 @@ import (
 
 const supported SupportedType = Links | ATime | BTime | CTime | UID | GID
 
+const userIDSource UserIDSourceType = UserIDSourceIsNumeric
+
 func (fs *fileStat) times() {
 	fs.atime = time.Unix(int64(fs.sys.Atimespec.Sec), int64(fs.sys.Atimespec.Nsec))
 	fs.btime = time.Unix(int64(fs.sys.Birthtimespec.Sec), int64(fs.sys.Birthtimespec.Nsec))
 	fs.ctime = time.Unix(int64(fs.sys.Ctimespec.Sec), int64(fs.sys.Ctimespec.Nsec))
 }
+
+func (fs *fileStat) BTime() time.Time { return fs.btime }