+++ stat_unix_stat.go
@@ -9,7 +9,10 @@ package compat
 
 import (
 	"os"
+	"os/user"
+	"strconv"
 	"syscall"
+	"time"
 )
 
 func stat(fi os.FileInfo, name string) (FileInfo, error) {
@@ -24,9 +27,43 @@ func stat(fi os.FileInfo, name string) (FileInfo, error) {
 	fs.partID = uint64(fs.sys.Dev) //nolint:gosec,unconvert,nolintlint // intentional int32 → uint64 conversion
 	fs.fileID = fs.sys.Ino
 	fs.links = uint64(fs.sys.Nlink) //nolint:gosec,unconvert,nolintlint // intentional int32 → uint64 conversion
-	fs.uid = uint64(fs.sys.Uid)
-	fs.gid = uint64(fs.sys.Gid)
+	fs.uid = int(fs.sys.Uid)
+	fs.gid = int(fs.sys.Gid)
+
 	fs.times()
 
 	return &fs, nil
+}
+
+func (fs *fileStat) CTime() time.Time { return fs.ctime }
+
+func (fs *fileStat) UID() int { return fs.uid }
+func (fs *fileStat) GID() int { return fs.gid }
+
+func (fs *fileStat) User() string {
+	if !fs.usered {
+		u, err := user.LookupId(strconv.Itoa(fs.uid))
+		if err != nil {
+			fs.err = err
+		} else {
+			fs.user = u.Username
+		}
+		fs.usered = true
+	}
+
+	return fs.user
+}
+
+func (fs *fileStat) Group() string {
+	if !fs.grouped {
+		g, err := user.LookupGroupId(strconv.Itoa(fs.gid))
+		if err != nil {
+			fs.err = err
+		} else {
+			fs.group = g.Name
+		}
+		fs.grouped = true
+	}
+
+	return fs.group
 }