+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased](https://github.com/rasa/compat/compare/v0.4.4...HEAD)
+## [Unreleased](https://github.com/rasa/compat/compare/v0.5.0...HEAD)
+
+### Added
+
+### Fixed
+
+### Changed
+
+## [0.5.0](https://github.com/rasa/compat/compare/v0.4.4...v0.5.0)
 
 ### Added